notes/day_19/requirements.txt
flask-cors==3.0.8
flask-restx==0.1.1
jsonref==0.2
jsonschema==3.2.0
pymongo==3.10.1
requests==2.23.0
